ref: renamed SideBarConfig.test to not have .test
By removing the .test extension from the dedicated SideBarConfig for test-cases jest does not consider this file a test and does executes all tests as expected.
