+++ tests/jest/root/SideBarConfig.ts
